% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longbeach}
\alias{longbeach}
\title{Long Beach Animal Shelter}
\format{
A data frame with 29787 rows and 23
variables.
}
\source{
\url{https://data.longbeach.gov/explore/dataset/animal-shelter-intakes-and-outcomes/information/}
}
\usage{
longbeach
}
\description{
A data set containing animals admitted to Long Beach Animal Shelter from
January 1, 2017 to January 1, 2025.
}
\keyword{datasets}
